LayoutTests/editing/pasteboard/copy-emoji-img-with-svg-source-expected.txt

This tests that emoji img elements with SVG sources are replaced with their Unicode alt text during copy serialization. This ensures clipboard HTML uses native Unicode emoji instead of external SVG references that may fail to load cross-origin.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS event.clipboardData.getData("text/html").includes("😌") is true
PASS event.clipboardData.getData("text/html").includes("emoji.svg") is false
PASS event.clipboardData.getData("text/html").includes("Hello") is true
PASS event.clipboardData.getData("text/html").includes("world") is true
PASS event.clipboardData.getData("text/plain").includes("😌") is true
PASS successfullyParsed is true

TEST COMPLETE
